+++ data-owner/src/main/java/de/unileipzig/dbs/pprl/service/dataowner/config/CsvDatasetConfig.java
@@ -20,7 +20,7 @@ public class CsvDatasetConfig {
 
     private String location;
 
-    private int datasetId;
+    private long datasetId;
 
     private String source;
   }
